Refactor courses feature and restructure appbar and footer components
- Updated AnimatedCoursesItem to use 'course' instead of 'service'.
- Removed CoursesPage and its associated logic, replacing it with a new structure for handling courses in desktop and mobile views.
- Created DeskTopCoursesPage and MobileExperiencesPage for better separation of concerns.
- Introduced new animated service item for displaying services with visibility detection.
- Reorganized appbar components, including language switch and theme toggle.
- Implemented a new footer structure with desktop and mobile variations.
- Added footer link items for social media integration.
