+++ lib/features/courses/widget/home_courses_list_mobile.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:gap/gap.dart';
 import 'package:my_prtofolio/features/courses/animation/animation_courses.dart';
 import 'package:my_prtofolio/features/courses/logic/cubit/courses_cubit.dart';
 import 'package:my_prtofolio/features/courses/logic/cubit/courses_state.dart';
-import 'package:my_prtofolio/features/courses/widget/courses_item.dart';
-import 'package:visibility_detector/visibility_detector.dart';
 
 class HomeCoursesListMobile extends StatelessWidget {
   HomeCoursesListMobile({super.key});
@@ -35,7 +32,7 @@ class HomeCoursesListMobile extends StatelessWidget {
                   itemBuilder:
                       (context, index) => Padding(
                         padding: const EdgeInsets.all(15.0),
-                        child: AnimatedCoursesItem(service: courses[index]),
+                        child: AnimatedCoursesItem(course: courses[index]),
                       ),
                 ),
               ),
