+++ lib/features/courses_page/presentation/courses_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:my_prtofolio/features/courses_page/widget/desktop_courses_page.dart';
+import 'package:my_prtofolio/features/courses_page/widget/mobile_courses_page.dart';
+import 'package:my_prtofolio/helper/extensions.dart';
+
+class CoursesPage extends StatelessWidget {
+  const CoursesPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return context.isDesktop || context.isTablet
+        ? const DeskTopCoursesPage()
+        : const MobileExperiencesPage();
+  }
+}
