+++ lib/features/home/presentation/blog/presentation/blog_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:my_prtofolio/features/home/presentation/appbar/widget/app_scafflod.dart';
+import 'package:my_prtofolio/shared/appbar/widget/app_scafflod.dart';
 
 class BlogPage extends StatelessWidget {
   const BlogPage({super.key});
