+++ lib/features/home/widget/hero_image.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:my_prtofolio/constants/assets.dart';
 
 class HeroImage extends StatelessWidget {
   const HeroImage({super.key});
@@ -9,7 +10,7 @@ class HeroImage extends StatelessWidget {
       aspectRatio: 0.6,
       child: ClipRRect(
         borderRadius: BorderRadius.all(Radius.circular(250)),
-        child: Image.asset('assets/images/me.jpg', fit: BoxFit.cover),
+        child: Image.asset(Assets.imagesMe, fit: BoxFit.cover),
       ),
     );
   }
