+++ lib/shared/appbar/drawer/small_menus.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:my_prtofolio/constant/app_menu_list.dart';
-import 'package:my_prtofolio/features/home/presentation/appbar/widget/large_app_bar_menu_item.dart';
+import 'package:my_prtofolio/shared/appbar/widget/large_app_bar_menu_item.dart';
 
-class LargeMenus extends StatelessWidget {
-  const LargeMenus({super.key});
+class SmallMenus extends StatelessWidget {
+  const SmallMenus({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Row(
+    return Column(
       children:
           AppMenuList.getItem(context)
               .map(
