+++ lib/shared/footer/widget/desktop_footer.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:my_prtofolio/features/home/presentation/appbar/widget/app_logo.dart';
-import 'package:my_prtofolio/features/home/presentation/appbar/drawer/large_menus.dart';
- import 'package:my_prtofolio/features/home/presentation/footer/widget/footer_links.dart';
+import 'package:my_prtofolio/shared/appbar/widget/app_logo.dart';
+import 'package:my_prtofolio/shared/appbar/drawer/large_menus.dart';
+ import 'package:my_prtofolio/shared/footer/widget/footer_links.dart';
 
 class DesktopFooter extends StatelessWidget {
   const DesktopFooter({super.key});
