+++ lib/shared/routes/app_route.dart
@@ -1,7 +1,8 @@
 import 'package:go_router/go_router.dart';
+import 'package:my_prtofolio/features/courses_page/presentation/courses_page.dart';
 import 'package:my_prtofolio/features/home/presentation/about/about_page.dart';
 import 'package:my_prtofolio/features/home/presentation/blog/presentation/blog_page.dart';
-import 'package:my_prtofolio/features/courses/presentation/courses_page.dart';
+import 'package:my_prtofolio/features/courses_page/widget/desktop_courses_page.dart';
 import 'package:my_prtofolio/features/home/presentation/home_page.dart';
 
 class Routes {
